StegaStamp/requirement_stega.txt
torchjpeg
einops
tqdm
numpy==1.26.4